copy and pastes the code from test.py to main.py, deletes test.py
